Finishing Touches
Tested the map, saw a bunch of lighting issues. Ongoing with making sure everywhere can be reached and works as provided.

Added a bunch of lamps to the left-side; and made it feel more swampish. Im getting very happy with the vibes.

Ambiance is working. No mob ambushes either. I did it.

Now we just need new music + Fog Effect